Allow bind9 deployments without zones
It can be useful to set up BIND9's named as a purely pass-through
recursive caching resolver, with no zones of its own to be authoritative
for.  Unfortunately, this is not allowed, due to the mechanics of how
zones (if supplied) are validated.

This commit fixes that oversight, allowing the checker to iterate over
<no zones> properly (i.e. never) and get on to the good stuff of running
named.

Fixes #19